requirements.txt
flask
psutil
pyautogui
pyttsx3
pyperclip
requests
Pillow
python-dotenv
pywin32
gtts
playsound==1.2.2